移除商品中的 活动库存

+++ application-global/src/main/java/com/onek/util/stock/RedisStockUtil.java
@@ -47,7 +47,7 @@ public class RedisStockUtil {
         return Integer.parseInt(currentStock);
     }
 
-    public static int getActStockBySkuAndActno(long sku, int actno){
+    public static int getActStockBySkuAndActno(long sku, long actno){
         return 0;
     }
